move generator install steps to install subcommand
